# api/openapi.yaml
openapi: 3.1.0
info:
  title: LegalMind API
  description: |
    API for LegalMind - Jurisdiction-specific legal template generation with paywall,
    guided questionnaire flow, and attorney referral system.

    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.

    ## Rate Limiting
    - Authenticated: 100 requests/minute
    - Unauthenticated: 10 requests/minute

    ## Errors
    Standard HTTP status codes. Error responses follow format:
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable message",
        "details": {}
      }
    }
    ```
  version: 1.0.0
  contact:
    name: LegalMind Support
    email: support@legalmind.tech
  license:
    name: Proprietary
servers:
  - url: https://api.legalmind.tech/v1
    description: Production
  - url: https://api-staging.legalmind.tech/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
...
paths:
...
  /referral:
    post:
      tags: [Referrals]
      summary: Get attorney referral matches
      description: |
        Returns 1-3 partner attorneys matched to template and jurisdiction.
        Generates unique discount token for tracking.
        Displayed immediately after successful generation.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [template_code, jurisdiction]
              properties:
                template_code:
                  type: string
                jurisdiction:
                  $ref: '#/components/schemas/JurisdictionCode'
                audit_id:
                  type: string
                  description: Link to generation audit log (optional)
            example:
              template_code: SAFE_PM_V1
              jurisdiction: UK
              audit_id: "aud_abc123xyz"
      responses:
        '200':
          description: Partners matched
          content:
            application/json:
              schema:
                type: object
                properties:
                  referral_id:
                    type: string
                    example: "ref_xyz789"
                  partners:
                    type: array
                    items:
                      $ref: '#/components/schemas/PartnerReferral'
                    minItems: 1
                    maxItems: 3
                  expires_at:
                    type: string
                    format: date-time
                    description: Discount token expiration
              example:
                referral_id: "ref_xyz789"
                partners:
                  - name: "Lex & Co Solicitors"
                    jurisdiction: "UK"
                    specializations: ["fundraising", "corporate"]
                    bio: "Leading UK startup law firm with 15+ years experience..."
                    photo_url: "https://cdn.legalmind.tech/partners/lex-co.jpg"
                    discount_token: "LM-UK-20-abc123"
                    discount_percentage: 20
                    booking_url: "https://lexco.com/book?ref=LM-UK-20-abc123"
                    rating: 4.8
                expires_at: "2025-10-29T23:59:59Z"
  /referral/{referral_id}/cta-clicked:
    post:
      tags: [Referrals]
      summary: Track CTA click
      description: Called when user clicks "Get Attorney Review" CTA
      security:
        - BearerAuth: []
      parameters:
        - name: referral_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                partner_id:
                  type: string
      responses:
        '204':
          description: Event tracked
  /referral/{referral_id}/booking-webhook:
    post:
      tags: [Referrals]
      summary: Partner booking webhook
      description: |
        **Internal webhook** called by partner systems when booking confirmed.
        Used for commission tracking.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: referral_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [partner_id, discount_token, booking_confirmed]
              properties:
                partner_id:
                  type: string
                discount_token:
                  type: string
                booking_confirmed:
                  type: boolean
                booking_value:
                  type: number
                  description: Booking value for commission calculation
      responses:
        '200':
          description: Webhook processed
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===== USERS =====
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from authentication flow

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Partner API key for webhooks
  schemas:
...
    JurisdictionCode:
      type: string
      enum:
        - GLOBAL-EN
        - UK
        - US-DE
        - DE
        - FR
        - CZ
      description: |
        Jurisdiction codes:
        - GLOBAL-EN: Global English baseline
        - UK: United Kingdom
        - US-DE: United States (Delaware)
        - DE: Germany
        - FR: France
        - CZ: Czech Republic
...
    PartnerReferral:
      type: object
      properties:
        partner_id:
          type: string
        name:
          type: string
        jurisdiction:
          $ref: '#/components/schemas/JurisdictionCode'
        specializations:
          type: array
          items:
            type: string
        bio:
          type: string
        photo_url:
          type: string
          format: uri
        discount_token:
          type: string
          description: Unique token for this referral
        discount_percentage:
          type: number
        booking_url:
          type: string
          format: uri
          description: Booking URL with discount token embedded
        rating:
          type: number
          minimum: 0
          maximum: 5
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: FORBIDDEN
            message:
              type: string
              example: Subscription required to access this resource
            details:
              type: object
  responses:
    Unauthorized:
      description: Unauthorized - missing or invalid auth token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...